backend/pyproject.toml
[project]
name = "pm-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.3",
    "channels>=4.2.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.2",
    "databricks-sdk>=0.102.0",
    "django>=6.0.4",
    "django-celery-results>=2.6.0",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.17.1",
    "djangorestframework-simplejwt==5.5.1",
    "langchain>=1.2.15",
    "langchain-community>=0.3.0",
    "langfuse>=4.0.6",
    "langgraph>=1.1.6",
    "litellm>=1.83.4",
    "mlflow>=3.11.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.0.0",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.1",
    "redis>=7.4.0",
    "ruff>=0.15.10",
    "smolagents>=1.24.0",
]
